decouple params into separate file
